fix: added numpy dtype to obs_converter()

+++ gyms/energym/energymGymEnv.py
@@ -287,11 +287,11 @@ class EnergymGymEnv(gym.Env):
         if self.normalize:
             for key in self.cont_obs:
                 obs[key] = np.array(2 * (observation[key] - self.obs_low[key]) /
-                                    (self.obs_high[key] - self.obs_low[key]) - 1).reshape(1,)
+                                    (self.obs_high[key] - self.obs_low[key]) - 1, dtype=np.float32).reshape(1,)
 
         elif self.discretize:
             for key in self.cont_obs:
-                obs[key] = np.digitize(observation[key], self.val_bins_obs[key])
+                obs[key] = np.digitize(observation[key], self.val_bins_obs[key], dtype=np.float32)
 
         # convert to ndarray
         # observation = np.array(list(observation.values()), dtype=np.float).reshape(len(observation.values()), )
